Refactor string functions
