+++ builtin-functions/kphp-light/functions.txt
@@ -197,11 +197,6 @@ function warning($message ::: string) ::: void;
 /** @kphp-no-return */
 function critical_error($message ::: string) ::: void;
 
-function debug_print_string($str ::: string) ::: void;
-
-function byte_to_int($str ::: string) ::: ?int;
-function int_to_byte($v ::: int) ::: ?string;
-
 /** @kphp-extern-func-info interruptible */
 function set_timer(int $timeout, callable():void $callback) ::: void;
